(WIP) Comment Flow

+++ lib/services/firebase_service.dart
@@ -1,6 +1,7 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:unifit/enums/firebase_collection_enum.dart';
 import 'package:unifit/interfaces/firebase_service_interface.dart';
+import 'package:unifit/utils/exception_handlers.dart';
 
 class FirebaseService implements IFirebaseService {
   /// Provides methods to interact with Firebase collections.
@@ -8,6 +9,9 @@ class FirebaseService implements IFirebaseService {
       .collection(FirebaseCollectionEnum.event.name);
   final CollectionReference _userCollection = IFirebaseService.firebaseInstance
       .collection(FirebaseCollectionEnum.user.name);
+  final CollectionReference _commentCollection = IFirebaseService
+      .firebaseInstance
+      .collection(FirebaseCollectionEnum.comment.name);
 
   static final FirebaseService instance = FirebaseService._();
 
@@ -18,14 +22,27 @@ class FirebaseService implements IFirebaseService {
     switch (collection) {
       case FirebaseCollectionEnum.event:
         return _eventCollection;
+      case FirebaseCollectionEnum.comment:
+        return _commentCollection;
       case FirebaseCollectionEnum.user:
         return _userCollection;
     }
   }
 
   @override
+  Future<DocumentReference> add({
+    required Map<String, dynamic> data,
+    required FirebaseCollectionEnum collection,
+  }) async {
+    return await getCollection(collection: collection).add(data);
+  }
+
+  @override
   Future<DocumentSnapshot<Object?>> get(
-      {required String id, required FirebaseCollectionEnum collection}) async {
+      {String? id, required FirebaseCollectionEnum collection}) async {
+    if (id == null) {
+      throw InternalApplicationError('Document ID wasn\'t provided');
+    }
     DocumentSnapshot document =
         await getCollection(collection: collection).doc(id).get();
     return document;
@@ -33,25 +50,60 @@ class FirebaseService implements IFirebaseService {
 
   @override
   Future<void> set(
-      {required String id,
-      value,
+      {String? id,
+      Map<String, dynamic>? data,
       required FirebaseCollectionEnum collection}) async {
-    await getCollection(collection: collection).doc(id).set(value);
+    if (id != null && data != null) {
+      await getCollection(collection: collection).doc(id).set(data);
+    }
   }
 
   @override
   Future<void> update(
-      {required String id,
-      value,
+      {String? id,
+      Map<String, dynamic>? data,
       required FirebaseCollectionEnum collection}) async {
-    await getCollection(collection: collection).doc(id).update(value);
+    if (id != null && data != null) {
+      await getCollection(collection: collection).doc(id).update(data);
+    }
   }
 
   @override
   Future<void> delete(
-      {required String id,
-      value,
-      required FirebaseCollectionEnum collection}) async {
-    await getCollection(collection: collection).doc(id).delete();
+      {String? id, required FirebaseCollectionEnum collection}) async {
+    if (id != null) {
+      await getCollection(collection: collection).doc(id).delete();
+    }
+  }
+
+  Future<Map<String, dynamic>?> getDocumentByMapValues(
+    Map<String, dynamic> values,
+    CollectionReference collectionReference,
+  ) async {
+    Query query = collectionReference;
+
+    // Construir a consulta com base nos valores do mapa
+    values.forEach((key, value) {
+      query = query.where(key, isEqualTo: value);
+    });
+
+    QuerySnapshot querySnapshot = await query.get();
+
+    if (querySnapshot.docs.isNotEmpty) {
+      for (QueryDocumentSnapshot documentSnapshot in querySnapshot.docs) {
+        if (values.entries.every((entry) {
+          final Map<String, dynamic>? data =
+              documentSnapshot.data() as Map<String, dynamic>?;
+
+          return data != null &&
+              data.containsKey(entry.key) &&
+              data[entry.key] == entry.value;
+        })) {
+          return documentSnapshot.data() as Map<String, dynamic>;
+        }
+      }
+    }
+
+    return null;
   }
 }
